fix: eliminate checkout model mixin place-order-mixin.js
This mixin was not properly implemented. It should have called the parent placeOrder implementation and modified the behavior. Instead it was completely replacing the placeOrder method. This prevented other mixins, for example the Recaptcha mixing, from functioning. Removing this mixin resolves those issues.

Looked at the default M2 behavior in the case where a payment method failed to save. That default behavior is to redirect to the shipping step (`#shipping`). Decided to stick with this default behavior in the case of 3DS2 instead of tryiing to stay on the payment step and display an error message.

Refs: #277